Use the public API
